+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Readme section: modified 'Run' button into a link
+- Hint extension for tile widget account
+
 ## [0.31.0] - 2024-11-22
 ### Added
 - Added Terms of Service link
-- Hint extension for tile widget account
 
 ## [0.30.0] - 2024-11-22
 ### Added